apps/api/celery_app.py
import os
from celery import Celery
from dotenv import load_dotenv

load_dotenv()

REDIS_URL = os.getenv("REDIS_URL", "redis://localhost:6379/0")

celery = Celery(
    "wincallem",
    broker=REDIS_URL,
    backend=REDIS_URL,
    include=["tasks"],
)

# Optional: Celery config (timeouts, serialization, etc.)
celery.conf.update(
    task_serializer="json",
    result_serializer="json",
    accept_content=["json"],
    task_track_started=True,
    broker_connection_retry_on_startup=True,
)
